tristate area in hexagon
This on is still a work in progress.  I was able to get CT in hexagon shapes.  YAY!  I tried with NY and it was upsidedown.  Just need to work out a few things.  I also think I found another way instead of using Illustrator to get the points.  I'm close.